Counting and Capping for Giant Spider Spawning

+++ code/game/objects/effects/spiders.dm
@@ -67,16 +67,18 @@
 	START_PROCESSING(SSobj, src)
 
 /obj/structure/spider/eggcluster/process()
-	amount_grown += rand(0,2)
-	if(amount_grown >= 100)
-		var/num = rand(3, 12)
-		for(var/i in 1 to num)
-			var/obj/structure/spider/spiderling/S = new /obj/structure/spider/spiderling(loc)
-			S.faction = faction.Copy()
-			S.master_commander = master_commander
-			if(player_spiders)
-				S.player_spiders = 1
-		qdel(src)
+	if(SSmobs.giant_spiders <= 30) //eggs gonna chill out until there is less spiders
+		amount_grown += rand(0,2)
+
+		if(amount_grown >= 100)
+			var/num = rand(3, 12)
+			for(var/i in 1 to num)
+				var/obj/structure/spider/spiderling/S = new /obj/structure/spider/spiderling(loc)
+				S.faction = faction.Copy()
+				S.master_commander = master_commander
+				if(player_spiders)
+					S.player_spiders = 1
+			qdel(src)
 
 /obj/structure/spider/spiderling
 	name = "spiderling"
@@ -168,23 +170,27 @@
 	if(isturf(loc))
 		amount_grown += rand(0,2)
 		if(amount_grown >= 100)
-			if(!grow_as)
-				grow_as = pick(typesof(/mob/living/simple_animal/hostile/poison/giant_spider))
-			var/mob/living/simple_animal/hostile/poison/giant_spider/S = new grow_as(loc)
-			S.faction = faction.Copy()
-			S.master_commander = master_commander
-			if(player_spiders && !selecting_player)
-				selecting_player = 1
-				spawn()
-					var/list/candidates = SSghost_spawns.poll_candidates("Do you want to play as a giant spider?", ROLE_GSPIDER, TRUE, source = S)
+			if (SSmobs.giant_spiders <= 40)
+				if(!grow_as)
+					grow_as = pick(typesof(/mob/living/simple_animal/hostile/poison/giant_spider))
+				var/mob/living/simple_animal/hostile/poison/giant_spider/S = new grow_as(loc)
+				SSmobs.giant_spiders++
+				S.faction = faction.Copy()
+				S.master_commander = master_commander
+				if(player_spiders && !selecting_player)
+					selecting_player = 1
+					spawn()
+						var/list/candidates = SSghost_spawns.poll_candidates("Do you want to play as a giant spider?", ROLE_GSPIDER, TRUE, source = S)
 
-					if(candidates.len)
-						var/mob/C = pick(candidates)
-						if(C)
-							S.key = C.key
-							if(S.master_commander)
-								to_chat(S, "<span class='biggerdanger'>You are a spider who is loyal to [S.master_commander], obey [S.master_commander]'s every order and assist [S.master_commander.p_them()] in completing [S.master_commander.p_their()] goals at any cost.</span>")
-			qdel(src)
+						if(candidates.len)
+							var/mob/C = pick(candidates)
+							if(C)
+								S.key = C.key
+								if(S.master_commander)
+									to_chat(S, "<span class='biggerdanger'>You are a spider who is loyal to [S.master_commander], obey [S.master_commander]'s every order and assist [S.master_commander.p_them()] in completing [S.master_commander.p_their()] goals at any cost.</span>")
+				qdel(src)
+			else
+				Destroy()
 
 /obj/structure/spider/spiderling/proc/random_skitter()
 	var/list/available_turfs = list()
